# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cafe & Restaurant Website
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SiteContent {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  heroImageUrl: string;
  aboutTitle: string;
  aboutText: string;
  aboutImageUrl: string;
  restaurantName: string;
  restaurantTagline: string;
  logoUrl: string;
  address: string;
  phone: string;
  email: string;
  openingHours: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  footerText: string;
  updatedAt: string;
}

export interface UpdateSiteContentBody {
  heroTitle?: string;
  heroSubtitle?: string;
  heroImageUrl?: string;
  aboutTitle?: string;
  aboutText?: string;
  aboutImageUrl?: string;
  restaurantName?: string;
  restaurantTagline?: string;
  logoUrl?: string;
  address?: string;
  phone?: string;
  email?: string;
  openingHours?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  footerText?: string;
}

export interface ThemeSettings {
  id: number;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  fontFamily: string;
  updatedAt: string;
}

export interface UpdateThemeBody {
  primaryColor?: string;
  secondaryColor?: string;
  accentColor?: string;
  fontFamily?: string;
}

export interface MenuCategory {
  id: number;
  name: string;
  description: string;
  iconName: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateMenuCategoryBody {
  name: string;
  description: string;
  iconName: string;
  sortOrder?: number;
}

export interface MenuItem {
  id: number;
  categoryId: number;
  name: string;
  description: string;
  price: string;
  imageUrl: string;
  isAvailable: boolean;
  isFeatured: boolean;
  tags: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  description: string;
  price: string;
  imageUrl: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
  tags?: string;
  sortOrder?: number;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  caption: string;
  altText: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryImageBody {
  imageUrl: string;
  caption: string;
  altText: string;
  sortOrder?: number;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface CreateInquiryBody {
  name: string;
  email: string;
  phone?: string;
  subject: string;
  message: string;
}

export type GetMenuItemsParams = {
  categoryId?: number;
};