# lib/api-zod/src/generated/types/createMenuCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cafe & Restaurant Website
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuCategoryBody {
  name: string;
  description: string;
  iconName: string;
  sortOrder?: number;
}