# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cafe & Restaurant Website
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  categoryId: number;
  name: string;
  description: string;
  price: string;
  imageUrl: string;
  isAvailable?: boolean;
  isFeatured?: boolean;
  tags?: string;
  sortOrder?: number;
}