# lib/api-zod/src/generated/types/siteContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Cafe & Restaurant Website
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContent {
  id: number;
  heroTitle: string;
  heroSubtitle: string;
  heroImageUrl: string;
  aboutTitle: string;
  aboutText: string;
  aboutImageUrl: string;
  restaurantName: string;
  restaurantTagline: string;
  logoUrl: string;
  address: string;
  phone: string;
  email: string;
  openingHours: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  footerText: string;
  updatedAt: string;
}